+++ NEWS.md
@@ -1,5 +1,15 @@
 Changes in version 1.6.6 (2017-11-10)
 
+- Improvements
+  * Package sticker colors are obtained from ColorBrewer
+  * babyGRAN packages will not include unnecessary man and vignettes dirs
+
+- Bugfixes
+  * Overall build stats will exclude suspended packages
+  * Handle cases when there is no package update risk during risk report creation
+
+Changes in version 1.6.6 (2017-11-10)
+
 - New Features
   * Parallelize creation of covr reports
   * Create a CRAN-style archive directory, location can be optionally specified in repo params